requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sse-starlette==2.1.3
boto3==1.39.3
python-dotenv==1.1.1
httpx==0.28.1
strands-agents==0.1.5
strands-agents-tools==0.1.3
ag-ui-protocol==0.1.15
